requirements.txt
numpy==2.2.6
pillow==12.0.0
librosa==0.11.0
matplotlib==3.10.7
soundfile==0.13.1
scipy==1.16.3
